Moved stuff around and added additional override for the support API
