+++ source/common.h
@@ -18,6 +18,5 @@ PyObject *aws_py_thread_join_all_managed(PyObject *self, PyObject *args);
 PyObject *aws_py_load_system_environment(PyObject *self, PyObject *args);
 PyObject *aws_py_is_env_ec2(PyObject *self, PyObject *args);
 PyObject *aws_py_get_ec2_instance_type(PyObject *self, PyObject *args);
-PyObject *aws_py_is_crt_s3_optimized_for_ec2_instance_type(PyObject *self, PyObject *args);
 
 #endif /* AWS_CRT_PYTHON_COMMON_H */
