+++ source/s3.h
@@ -7,6 +7,8 @@
 
 #include "module.h"
 
+PyObject *aws_py_s3_is_crt_s3_optimized_for_system_env(PyObject *self, PyObject *args);
+
 PyObject *aws_py_s3_client_new(PyObject *self, PyObject *args);
 PyObject *aws_py_s3_client_make_meta_request(PyObject *self, PyObject *args);
